requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiogram.contrib==1.1.4
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.2.0
packaging==24.2
propcache==0.3.1
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
toml==0.10.2
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
